requirements.txt
blis==0.7.5
catalogue==2.0.6
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.7
click==7.1.2
colorama==0.4.4
cymem==2.0.6
DAWG-Python==0.7.2
docopt==0.6.2
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.11.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
install==1.3.4
intervaltree==3.1.0
ipymarkup==0.9.0
Jinja2==3.0.3
joblib==1.1.0
langcodes==3.3.0
MarkupSafe==2.0.1
murmurhash==1.0.6
natasha==1.4.0
navec==0.10.0
numpy==1.21.4
opencv-python==4.5.4.58
packaging==21.2
pandas==1.3.4
pathy==0.6.1
Pillow==8.4.0
preshed==3.0.6
pydantic==1.8.2
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pyparsing==2.4.7
PyQt5==5.15.6
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyqt5-tools==5.15.4.3.2
pytesseract==0.3.8
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
razdel==0.5.0
requests==2.26.0
rfc3986==1.5.0
ru-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/ru_core_news_sm-3.2.0/ru_core_news_sm-3.2.0-py3-none-any.whl
scikit-learn==1.0.1
scipy==1.7.2
six==1.16.0
sklearn==0.0
slovnet==0.5.0
smart-open==5.2.1
sniffio==1.2.0
sortedcontainers==2.4.0
spacy==3.2.0
spacy-legacy==3.0.8
spacy-loggers==1.0.1
srsly==2.4.2
thinc==8.0.13
threadpoolctl==3.0.0
tqdm==4.62.3
typer==0.4.0
typing-extensions==3.10.0.2
urllib3==1.26.7
wasabi==0.8.2
yargy==0.15.0
